Student Study Portal Version 1
Created a website that allows students to take down notes and homework. Wikipedia, YouTube and books (Will be added on Version 2) are used to bookmark to aid students on which one is the most important thing you need to do for that day. Enquires on the problems for the website to their e-mails is also added into this project.